Refactor percona/controller
