+++ percona/controller/pgcluster/suite_test.go
@@ -1,4 +1,4 @@
-package controllers
+package pgcluster
 
 import (
 	"os"
