+++ percona/controller/utils.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 import (
 	batchv1 "k8s.io/api/batch/v1"
@@ -7,7 +7,7 @@ import (
 
 // jobCompleted returns "true" if the Job provided completed successfully.  Otherwise it returns
 // "false".
-func jobCompleted(job *batchv1.Job) bool {
+func JobCompleted(job *batchv1.Job) bool {
 	conditions := job.Status.Conditions
 	for i := range conditions {
 		if conditions[i].Type == batchv1.JobComplete {
@@ -18,7 +18,7 @@ func jobCompleted(job *batchv1.Job) bool {
 }
 
 // jobFailed returns "true" if the Job provided has failed.  Otherwise it returns "false".
-func jobFailed(job *batchv1.Job) bool {
+func JobFailed(job *batchv1.Job) bool {
 	conditions := job.Status.Conditions
 	for i := range conditions {
 		if conditions[i].Type == batchv1.JobFailed {
